+++ src/Application/Query/Order/Export/OrderViewQueryHandler.php
@@ -131,7 +131,8 @@ class OrderViewQueryHandler
                     $query->event->getMode(),
                     $query->event->getCountry(),
                     $billingInfoView->countryCode,
-                    $billingInfoView->vatNumber
+                    $billingInfoView->vatNumber,
+                    $sheet
                 )
             )
         );